requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 数据处理
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0
albumentations>=1.3.0

# 配置和日志
PyYAML>=6.0
tensorboard>=2.10.0
tqdm>=4.64.0

# 科学计算
scipy>=1.9.0
scikit-learn>=1.1.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 工具库
pandas>=1.4.0
requests>=2.28.0

# 开发工具
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# 其他
timm>=0.6.0
thop>=0.1.0 